+++ tests/testthat/test-1-autoplot.R
@@ -171,7 +171,7 @@ test_that("autplot blockCV shows correct blocks", {
   task = test_make_blockCV_test_task()
 
   set.seed(42)
-  rsmp <- rsmp("spcv_block",
+  rsmp = rsmp("spcv_block",
     folds = 5,
     rows = 3,
     cols = 4)
@@ -191,7 +191,7 @@ test_that("autplot blockCV shows correct blocks for repeated_cv", {
   task = test_make_blockCV_test_task()
 
   set.seed(42)
-  rsmp <- rsmp("repeated_spcv_block",
+  rsmp = rsmp("repeated_spcv_block",
     folds = 5,
     repeats = 2,
     rows = 3,
@@ -446,4 +446,45 @@ test_that("plot() works for 'spcv_disc'", {
 
   vdiffr::expect_doppelganger("RepSpCVDisc - Fold 1-2, Rep 2", p5)
   vdiffr::expect_doppelganger("RepSpCVDisc - Fold 1, Rep 2", p6)
+})
+
+# spcv_tiles --------------------------------------------------------------------
+
+test_that("plot() works for 'spcv_tiles'", {
+  skip_if_not_installed("vdiffr")
+
+  # for some reason linux and windows svgs differ
+  skip_on_os("linux")
+  skip_on_os("windows")
+
+  set.seed(42)
+
+  task = tsk("ecuador")
+  rsp = rsmp("repeated_spcv_tiles",
+    nsplit = c(4L, 3L), min_n = 50, reassign = FALSE,
+    repeats = 2)
+  rsp$instantiate(task)
+
+  p1 = autoplot(rsp, task = task, crs = 4326)
+  p2 = autoplot(rsp, task, 1, crs = 4326)
+  # plot() would force image printing here
+  p3 = autoplot(rsp, task, c(1, 2), crs = 4326)
+
+  p4 = autoplot(rsp, task, 1, crs = 4326, show_omitted = TRUE)
+
+  expect_true(is.ggplot(p1))
+  expect_true(is.ggplot(p2))
+  expect_list(p3)
+  expect_true(is.ggplot(p4))
+
+  p5 = autoplot(rsp, task, crs = 4326, c(1, 2), repeats_id = 2)
+  p6 = autoplot(rsp, task, crs = 4326, fold_id = 1, repeats_id = 2)
+
+  vdiffr::expect_doppelganger("SpCVTiles all test sets", p1)
+  vdiffr::expect_doppelganger("SpCVTiles Fold 1", p2)
+  vdiffr::expect_doppelganger("SpCVTiles Fold 1-2", p3)
+  vdiffr::expect_doppelganger("SpCVTiles show_omitted", p4)
+
+  vdiffr::expect_doppelganger("RepSpCVTiles - Fold 1-2, Rep 2", p5)
+  vdiffr::expect_doppelganger("RepSpCVTiles - Fold 1, Rep 2", p6)
 })